% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Sum up two variables}
\usage{
add(a, b, plot = FALSE)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{plot}{Makes plot if TRUE}
}
\value{
Returns a numeric that is the sum of a and b.
}
\description{
What is the sum of a and b?
}
\examples{
add(7,2)
add(4,16,plot=TRUE)

}
